moved eventDetailsTemplate.ts to components from assets folder
