+++ frontend/src/components/eventDetailsTemplate.ts
